Create USER STORIES
USER STORIES created to get some ideas going, but primary objective is to explain to developers applications for subject matters they may not be familiar with. The purpose of these user stories are to help developers in coding and configuring. 
The secondary purpose of the USER STORIES are to get others thinking of unique applications for hardware and blockchain and spur creativity in order to use technology and IoT in a responsible and open source fashion.